fix mistakes.

+++ services/movie_session.py
@@ -8,7 +8,9 @@ from db.models import MovieSession
 
 @transaction.atomic
 def create_movie_session(
-    movie_show_time: datetime, movie_id: int, cinema_hall_id: int
+        movie_show_time: datetime,
+        movie_id: int,
+        cinema_hall_id: int
 ) -> None:
     MovieSession.objects.create(
         show_time=movie_show_time,
@@ -36,10 +38,10 @@ def get_movie_session_by_id(movie_session_id: int) -> MovieSession | None:
 
 @transaction.atomic
 def update_movie_session(
-    session_id: int,
-    show_time: datetime = None,
-    movie_id: int = None,
-    cinema_hall_id: int = None,
+        session_id: int,
+        show_time: datetime = None,
+        movie_id: int = None,
+        cinema_hall_id: int = None,
 ) -> None:
     movie_session = get_movie_session_by_id(session_id)
     if not movie_session:
@@ -57,4 +59,6 @@ def update_movie_session(
 
 def delete_movie_session_by_id(session_id: int) -> None:
     movie_session = get_movie_session_by_id(session_id)
+    if not movie_session:
+        return
     movie_session.delete()
